+++ src/views/blocks/LayoutBlock.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ * View file for block: LayoutBlock 
+ *
+ * File has been created with `block/create` command. 
+ *
+ * @param $this->placeholderValue('content');
+ *
+ * @var $this \luya\cms\base\PhpBlockView
+ */
+?>
+
+<div class="row">
+    <?= $this->placeholderValue('content') ?>
+</div>